chore: improve performance of the session tests
